services/encryption/requirements.txt
fastapi
uvicorn[standard]
cryptography
pycryptodome
requests
python-multipart
minio